requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.2.0
auto-py-to-exe==2.10.1
bidict==0.21.4
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colorlog==6.5.0
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
Eel==0.12.4
et-xmlfile==1.1.0
excel2json-3==0.1.6
fake-useragent==0.1.11
Flask==2.0.2
Flask-Dropzone==1.6.0
Flask-FontAwesome==0.1.5
Flask-SocketIO==5.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.12.0
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.3.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
ntplib==0.4.0
numpy==1.21.2
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.3
pefile==2021.9.3
Pillow==8.4.0
PyAutoGUI==0.9.53
pycparser==2.20
pyfiglet==0.8.post1
PyGetWindow==0.0.9
pygtail==0.11.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.2.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.1
python-engineio==4.3.0
python-socketio==5.4.1
pytweening==1.0.4
pytz==2021.3
pywin32-ctypes==0.2.0
random-user-agent==1.0.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
ua-parser==0.10.0
undetected-chromedriver==3.0.3
urllib3==1.26.7
user-agents==2.2.0
webdriver-manager==3.5.0
websocket-client==1.1.0
websockets==10.0
Werkzeug==2.0.2
wget==3.2
whichcraft==0.6.1
wsproto==1.0.0
WTForms==2.3.3
xlrd==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
